+++ libosmscout-import/src/osmscout/import/GenAreaWayIndex.cpp
@@ -58,7 +58,7 @@ namespace osmscout {
                                                   TypeData& typeData,
                                                   const CoordCountMap& cellFillCount)
   {
-    typeData.indexLevel=level;
+    typeData.indexLevel=(uint32_t)level;
     typeData.indexCells=cellFillCount.size();
     typeData.indexEntries=0;
 
@@ -348,13 +348,13 @@ namespace osmscout {
         // by the way
         // Renormated coordinate space (everything is >=0)
         //
-        size_t minxc=(size_t)floor((minLon+180.0)/cellWidth);
-        size_t maxxc=(size_t)floor((maxLon+180.0)/cellWidth);
-        size_t minyc=(size_t)floor((minLat+90.0)/cellHeight);
-        size_t maxyc=(size_t)floor((maxLat+90.0)/cellHeight);
+        uint32_t minxc=(uint32_t)floor((minLon+180.0)/cellWidth);
+        uint32_t maxxc=(uint32_t)floor((maxLon+180.0)/cellWidth);
+        uint32_t minyc=(uint32_t)floor((minLat+90.0)/cellHeight);
+        uint32_t maxyc=(uint32_t)floor((maxLat+90.0)/cellHeight);
 
-        for (size_t y=minyc; y<=maxyc; y++) {
-          for (size_t x=minxc; x<=maxxc; x++) {
+        for (uint32_t y=minyc; y<=maxyc; y++) {
+          for (uint32_t x=minxc; x<=maxxc; x++) {
             cellFillCount[way.GetType()][Pixel(x,y)]++;
           }
         }
@@ -449,13 +449,13 @@ namespace osmscout {
         // by the way
         // Renormated coordinate space (everything is >=0)
         //
-        size_t minxc=(size_t)floor((minLon+180.0)/cellWidth);
-        size_t maxxc=(size_t)floor((maxLon+180.0)/cellWidth);
-        size_t minyc=(size_t)floor((minLat+90.0)/cellHeight);
-        size_t maxyc=(size_t)floor((maxLat+90.0)/cellHeight);
+        uint32_t minxc=(uint32_t)floor((minLon+180.0)/cellWidth);
+        uint32_t maxxc=(uint32_t)floor((maxLon+180.0)/cellWidth);
+        uint32_t minyc=(uint32_t)floor((minLat+90.0)/cellHeight);
+        uint32_t maxyc=(uint32_t)floor((maxLat+90.0)/cellHeight);
 
-        for (size_t y=minyc; y<=maxyc; y++) {
-          for (size_t x=minxc; x<=maxxc; x++) {
+        for (uint32_t y=minyc; y<=maxyc; y++) {
+          for (uint32_t x=minxc; x<=maxxc; x++) {
             cellFillCount[rel.GetType()][Pixel(x,y)]++;
           }
         }
@@ -621,13 +621,13 @@ namespace osmscout {
         // by the way
         // Renormated coordinate space (everything is >=0)
         //
-        size_t minxc=(size_t)floor((minLon+180.0)/cellWidth);
-        size_t maxxc=(size_t)floor((maxLon+180.0)/cellWidth);
-        size_t minyc=(size_t)floor((minLat+90.0)/cellHeight);
-        size_t maxyc=(size_t)floor((maxLat+90.0)/cellHeight);
+        uint32_t minxc=(uint32_t)floor((minLon+180.0)/cellWidth);
+        uint32_t maxxc=(uint32_t)floor((maxLon+180.0)/cellWidth);
+        uint32_t minyc=(uint32_t)floor((minLat+90.0)/cellHeight);
+        uint32_t maxyc=(uint32_t)floor((maxLat+90.0)/cellHeight);
 
-        for (size_t y=minyc; y<=maxyc; y++) {
-          for (size_t x=minxc; x<=maxxc; x++) {
+        for (uint32_t y=minyc; y<=maxyc; y++) {
+          for (uint32_t x=minxc; x<=maxxc; x++) {
             typeCellOffsets[way.GetType()][Pixel(x,y)].push_back(offset);
           }
         }
@@ -713,13 +713,13 @@ namespace osmscout {
         // by the way
         // Renormated coordinate space (everything is >=0)
         //
-        size_t minxc=(size_t)floor((minLon+180.0)/cellWidth);
-        size_t maxxc=(size_t)floor((maxLon+180.0)/cellWidth);
-        size_t minyc=(size_t)floor((minLat+90.0)/cellHeight);
-        size_t maxyc=(size_t)floor((maxLat+90.0)/cellHeight);
+        uint32_t minxc=(uint32_t)floor((minLon+180.0)/cellWidth);
+        uint32_t maxxc=(uint32_t)floor((maxLon+180.0)/cellWidth);
+        uint32_t minyc=(uint32_t)floor((minLat+90.0)/cellHeight);
+        uint32_t maxyc=(uint32_t)floor((maxLat+90.0)/cellHeight);
 
-        for (size_t y=minyc; y<=maxyc; y++) {
-          for (size_t x=minxc; x<=maxxc; x++) {
+        for (uint32_t y=minyc; y<=maxyc; y++) {
+          for (uint32_t x=minxc; x<=maxxc; x++) {
             typeCellOffsets[rel.GetType()][Pixel(x,y)].push_back(offset);
           }
         }
